only use ratings.js to download site data, port old CSV script to a separate place
